Fixed bug in pll_calc that did not allow tuning over 125 MHz, fixed corr. EEPROM

+++ si5351.h
@@ -196,8 +196,8 @@
  */
 
 # define do_div(n,base) ({                                      \
-        uint32_t __base = (base);                               \
-        uint32_t __rem;                                         \
+        uint64_t __base = (base);                               \
+        uint64_t __rem;                                         \
         __rem = ((uint64_t)(n)) % __base;                       \
         (n) = ((uint64_t)(n)) / __base;                         \
         __rem;                                                  \
